requirements.txt
timm==0.9.16
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
streamlit>=1.32.0
streamlit-webrtc>=0.47.0
aiortc>=1.8.0
av>=11.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
roboflow>=1.1.9